backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# Google Cloud Services
google-cloud-aiplatform>=1.50.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
google-cloud-translate>=3.16.0

# Security & Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0
